all: Add AccumulatorProof to Block
This also forced my hand regarding some p2p and chain code: previously
we had functions like AddTransactions and messages like MsgTransactions,
which were supposed to provide the "other half" of a full block (after
already having received the header). But now that blocks *also* have an
AccumulatorProof, it seemed prudent to quit trying to be super-minimal
and just deal in Blocks instead, even though this means duplicating the
header in most cases.

The MerkleProof field of Output is in sort of a limbo state now. It's
still used when broadcasting txns, so the wallet and txpool still need
to keep it updated, but once the full block has been assembled, things
get a little weird. The MerkleProof data is redundant with the block's
AccumulatorProof, but it's still really useful to have around. So
currently, we omit the individual MerkleProofs when serializing blocks,
but we reconstistute them for spent outputs and new outputs. At some
later point, I'll probably remove the MerkleProof field from Output
and introduce a OutputWithProof type for clients that need to maintain
outputs, so that there's no confusion about when the MerkleProof field
is "live" or not.