p2p: Detect invalid fields in checkpoint header
Header.ID() does not hash all of the fields in the Header. This means
that an attacker can modify the un-hashed fields and trick you into
accepting an invalid Header. To combat this, we need to manually check
the un-hashed fields. Normally, ValidateBlock will handle this, but in
DownloadCheckpoint we need to do it ourselves.

It is a little concerning how easy this is to miss -- I mean, clearly
I missed it, and I designed it! Perhaps ID() should hash all fields.