all: Partially revert c53e7f5d
Changed my mind on this. I started getting really multiproof-happy;
adding not only block-level multiproofs, but transaction-level ones
as well (for broadcasting txns). The end goal was to remove the
MerkleProof field from the Output struct entirely. But the deeper I
got into it, the less sense that made. It adds significant complexity
to a lot of code and creates an annoying distinction between "block"
transactions and "standalone" transactions. It also breaks the whole
conceptual model of "each output carries a proof of its validity." So
finally I decided to toss it all and go back to the old approach. Now,
multiproofs only exist as a size optimization, meaning they should
only be relevant when encoding (e.g. in p2p protocols and when writing
to disk). We transparently compress+expand multiproofs where needed,
so 99% of the time you can assume that each output has its own proof.

There are two downsides to this that I can think of. First, compressing
and expanding proofs requires CPU. The worst offender here is that we
sometimes re-compress proofs that were previously expanded. I don't
think this is a big deal -- hashing is cheap -- but maybe later we can
keep the proofs around instead of recomputing them. Second, expanding
the proofs means a bunch of redundant proof data will exist in memory.
Again, probably not a big deal -- memory is cheap -- but maybe there's
some attack surface here.

20% chance I end up changing my mind again (or being convinced by
someone smarter) and we go back to first-class multiproofs. We'll see!